+++ geo_rdm_records/__init__.py
@@ -9,5 +9,5 @@
 
 from .ext import GEORDMRecords
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"
 __all__ = ("__version__", "GEORDMRecords")