+++ geo_rdm_records/modules/requests/__init__.py
@@ -5,4 +5,4 @@
 # geo-rdm-records is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 
-"""GEO RDM Records CMS services module."""
+"""GEO RDM Records Auxiliary Requests module."""