+++ geo_rdm_records/modules/requests/services/config.py
@@ -5,24 +5,19 @@
 # geo-rdm-records is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 
-"""GEO RDM Records CMS service config module."""
+"""GEO RDM Records Requests service config module."""
 
 from invenio_records_resources.services.base.config import (
     ConfiguratorMixin,
-    FromConfig,
     ServiceConfig,
 )
 
 from geo_rdm_records.base.services.permissions import BaseGEOPermissionPolicy
 
 
-class CMSServiceConfig(ServiceConfig, ConfiguratorMixin):
-    """CMS Service config."""
+class AssistanceRequestsServiceConfig(ServiceConfig, ConfiguratorMixin):
+    """Assistance Requests for Packages Service config."""
 
     # Common configuration
-    service_id = "cms"
+    service_id = "assistance-requests"
     permission_policy_cls = BaseGEOPermissionPolicy
-
-    # CMS configuration
-    cms_api_token = FromConfig("GEO_RDM_RECORDS_CMS_API_TOKEN", import_string=False)
-    cms_api_address = FromConfig("GEO_RDM_RECORDS_CMS_API_ADDRESS", import_string=False)